Add createApolloServer function.
