+++ .changeset/2d1d6ed7/changes.json
@@ -1,0 +1,1 @@
+{ "releases": [{ "name": "@keystone-alpha/server", "type": "patch" }], "dependents": [] }
