+++ .changeset/2d1d6ed7/changes.md
@@ -1,0 +1,1 @@
+- Factor out a `createApolloServer` function.
